backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
django-extensions
djangorestframework-simplejwt
drf-yasg

psycopg2-binary
django-redis
celery
redis

channels
channels-redis
daphne

web3
eth-keys
eth-typing

Pillow
cloudinary
python-magic

requests
python-dotenv
pytz==2023.3.post1

python-socketio
python-engineio
aiohttp
